Show gameplay when loading the skin editor from the main menu
